test: extract shared Logs test fixtures into logFixtures.ts
Phase D1: MockLogEntry type, createMockLogEntry, and createMockLogs
were pasted identically into all seven Logs test files (~73 lines each).
Move them verbatim to web/src/test/logFixtures.ts and replace local
copies with imports.
