+++ web/src/pages/__tests__/Logs.updates-api.test.tsx
@@ -85,80 +85,7 @@ vi.mock("../../components/VirtualLogTable", () => ({
 	),
 }));
 
-// Factory functions for creating mock log entries
-interface MockLogEntry {
-	id: string;
-	created_at: string;
-	request_hash: string;
-	model_id: string;
-	provider_name: string;
-	status_code: number;
-	tokens_prompt: number;
-	tokens_completion: number;
-	tokens_per_second: number;
-	tokens_completion_reasoning: number;
-	ttft_ms: number;
-	response_header_ms: number;
-	duration_ms: number;
-	proxy_overhead_ms: number;
-	state: "completed" | "pending" | "streaming";
-	error_message: string;
-	parse_ms: number;
-	failover_lookup_ms: number;
-	model_lookup_ms: number;
-	provider_lookup_ms: number;
-	key_decrypt_ms: number;
-	dial_ms: number;
-	virtual_key_deleted: boolean;
-	virtual_key_id: string;
-	virtual_key_name?: string;
-}
-
-function createMockLogEntry(
-	overrides: Partial<MockLogEntry> = {},
-): MockLogEntry {
-	const defaultEntry: MockLogEntry = {
-		id: "log-001",
-		created_at: "2026-05-11T10:00:00Z",
-		request_hash: "abc123",
-		model_id: "test-model",
-		provider_name: "Test",
-		status_code: 200,
-		tokens_prompt: 0,
-		tokens_completion: 0,
-		tokens_per_second: 0,
-		tokens_completion_reasoning: 0,
-		ttft_ms: 0,
-		response_header_ms: 0,
-		duration_ms: 0,
-		proxy_overhead_ms: 0,
-		state: "completed",
-		error_message: "",
-		parse_ms: 0,
-		failover_lookup_ms: 0,
-		model_lookup_ms: 0,
-		provider_lookup_ms: 0,
-		key_decrypt_ms: 0,
-		dial_ms: 0,
-		virtual_key_deleted: false,
-		virtual_key_id: "",
-	};
-	return { ...defaultEntry, ...overrides };
-}
-
-function createMockLogs(
-	entries: MockLogEntry[],
-	total?: number,
-	page: number = 1,
-	perPage: number = 25,
-) {
-	return {
-		entries,
-		total: total ?? entries.length,
-		page,
-		per_page: perPage,
-	};
-}
+import { createMockLogEntry, createMockLogs } from "../../test/logFixtures";
 
 describe("Logs", () => {
 	beforeEach(() => {
